+++ app/src/main/res/layout/bookinfo_review_recyclerview.xml
@@ -1,49 +1,83 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.cardview.widget.CardView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:orientation="vertical">
+    app:cardCornerRadius="10dp"
+    app:cardElevation="4dp">
 
-    <TextView
-        android:id="@+id/tvBookInfoRv1"
-        android:layout_width="wrap_content"
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:ellipsize="end"
-        android:maxLines="3"
-        android:text="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
-        android:textSize="14sp" />
+        android:padding="20dp">
 
-    <!-- 별점 및 리뷰 수 -->
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="end"
-        android:gravity="center_vertical"
-        android:orientation="horizontal">
-
-        <ImageView
-            android:layout_width="16dp"
-            android:layout_height="16dp"
-            android:src="@drawable/baseline_star_24" />
-
+        <!-- 리뷰 텍스트 -->
         <TextView
-            android:id="@+id/tvBookInfoRvRate1"
+            android:id="@+id/reviewText"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:text="노벨상 수상 소식 이후로, 읽었던 때의 기억을 다시금 반추하며 리뷰를 작성해 본다..."
+            android:textSize="14sp"
+            android:textColor="@color/black"
+            android:lineSpacingExtra="4dp"
+            android:maxLines="3"
+            android:ellipsize="end"
+            android:fontFamily="@font/pretendard_variable"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintBottom_toTopOf="@+id/ratingBar"/>
+
+        <!-- 별점 (RatingBar) -->
+
+
+        <!-- 좋아요 아이콘 -->
+        <ImageView
+            android:id="@+id/likeIcon"
+            android:layout_width="20dp"
+            android:layout_height="20dp"
+            android:src="@drawable/recommend"
+            app:tint="@color/gray1"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toStartOf="@id/likeCount"
+            app:layout_constraintTop_toBottomOf="@id/reviewText"
+            android:layout_marginRight="4dp"/>
+
+        <!-- 좋아요 개수 -->
+        <TextView
+            android:id="@+id/likeCount"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="0.0"
-            android:textSize="14sp" />
+            android:text="189"
+            android:textColor="#B8B8B8"
+            android:textSize="10sp"
+            android:textStyle="normal"
+            android:fontWeight="400"
+            android:fontFamily="@font/pretendard_variable"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/reviewText"/>
 
-        <ImageView
-            android:layout_width="16dp"
-            android:layout_height="match_parent"
-            android:src="@drawable/baseline_thumb_up_24" />
+        <me.zhanghai.android.materialratingbar.MaterialRatingBar
+            android:id="@+id/ratingBar"
+            android:layout_width="58dp"
+            android:layout_height="10dp"
+            android:numStars="5"
+            android:rating="3"
+            android:stepSize="0.5"
+            android:isIndicator="true"
+            app:mrb_progressTint="@color/Primary"
+            app:mrb_progressBackgroundTint="@color/gray1"
+            app:mrb_secondaryProgressTint="@color/gray1"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="parent" />
+<!--            app:mrb_starCount="5"-->
+<!--            app:mrb_starSize="10dp"-->
+<!--            app:mrb_starPadding="2dp"-->
+<!--            app:mrb_touchable="false"-->
 
-        <TextView
-            android:id="@+id/tvBookInfoRvRecommend1"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="000"
-            android:textSize="14sp" />
-    </LinearLayout>
 
-</LinearLayout>
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+</androidx.cardview.widget.CardView>
